# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CineClub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OperationResult {
  success: boolean;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
  watchedCount?: number;
  reviewCount?: number;
  friendCount?: number;
}

export interface UserProfileUpdate {
  displayName?: string;
  username?: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
}

export interface Genre {
  id: number;
  name: string;
}

export interface GenreListResponse {
  genres: Genre[];
}

export interface MovieSummary {
  tmdbId: number;
  title: string;
  /** @nullable */
  posterPath: string | null;
  /** @nullable */
  backdropPath?: string | null;
  /** @nullable */
  releaseYear: number | null;
  voteAverage: number;
  popularity: number;
  genreIds?: number[];
  overview?: string;
}

export interface MovieListResponse {
  results: MovieSummary[];
  totalPages: number;
  page: number;
}

export interface CastMember {
  id: number;
  name: string;
  character: string;
  /** @nullable */
  profilePath?: string | null;
}

export interface MovieDetail {
  tmdbId: number;
  title: string;
  overview: string;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  /** @nullable */
  releaseYear: number | null;
  voteAverage: number;
  popularity: number;
  /** @nullable */
  runtime: number | null;
  genres: Genre[];
  cast?: CastMember[];
  /** @nullable */
  trailerKey?: string | null;
  /** @nullable */
  dominantColor?: string | null;
}

export interface MovieUserStatus {
  watchlistIds: number[];
  hasReview: boolean;
  /** @nullable */
  rating?: number | null;
}

export interface Watchlist {
  id: number;
  userId: string;
  name: string;
  isDefault: boolean;
  movieCount: number;
  createdAt: string;
}

export interface WatchlistInput {
  name: string;
}

export interface WatchlistUpdate {
  name: string;
}

export interface WatchlistMovie {
  id: number;
  watchlistId: number;
  tmdbId: number;
  title: string;
  /** @nullable */
  posterPath: string | null;
  /** @nullable */
  releaseYear?: number | null;
  voteAverage?: number;
  addedAt: string;
}

export interface WatchlistMovieInput {
  tmdbId: number;
  title: string;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  releaseYear?: number | null;
  voteAverage?: number;
}

export interface Review {
  id: number;
  userId: string;
  tmdbId: number;
  /** @nullable */
  movieTitle?: string | null;
  /** @nullable */
  moviePosterPath?: string | null;
  rating: number;
  content: string;
  isSpoiler: boolean;
  createdAt: string;
}

export interface ReviewWithUser {
  id: number;
  userId: string;
  tmdbId: number;
  rating: number;
  content: string;
  isSpoiler: boolean;
  createdAt: string;
  user: UserProfile;
}

export interface ReviewInput {
  tmdbId: number;
  movieTitle?: string;
  /** @nullable */
  moviePosterPath?: string | null;
  rating: number;
  content: string;
  isSpoiler: boolean;
}

export interface Friendship {
  id: number;
  requesterId: string;
  receiverId: string;
  status: string;
  createdAt: string;
  otherUser?: UserProfile;
}

export interface FriendRequestInput {
  targetUsername: string;
}

export interface FriendMovieActivity {
  user: UserProfile;
  /** @nullable */
  rating: number | null;
  /** @nullable */
  review?: string | null;
  isSpoiler?: boolean;
}

export interface Recommendation {
  id: number;
  fromUserId: string;
  toUserId: string;
  tmdbId: number;
  /** @nullable */
  message?: string | null;
  seen: boolean;
  createdAt: string;
}

export interface RecommendationWithMovie {
  id: number;
  fromUserId: string;
  toUserId: string;
  tmdbId: number;
  /** @nullable */
  message?: string | null;
  seen: boolean;
  createdAt: string;
  movie: MovieSummary;
  fromUser: UserProfile;
}

export interface RecommendationInput {
  toUserId: string;
  tmdbId: number;
  movieTitle?: string;
  /** @nullable */
  moviePosterPath?: string | null;
  /** @nullable */
  message?: string | null;
}

export type NotificationData = { [key: string]: unknown };

export interface Notification {
  id: number;
  userId: string;
  type: string;
  data: NotificationData;
  read: boolean;
  createdAt: string;
}

export type SearchMoviesParams = {
q: string;
page?: number;
};

export type GetPopularMoviesParams = {
page?: number;
};

export type GetRandomPickParams = {
watchlistId?: number;
genreId?: number;
};