# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CineClub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /movies/search:
    get:
      operationId: searchMovies
      tags: [movies]
      summary: Search movies via TMDB
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListResponse"
  /movies/popular:
    get:
      operationId: getPopularMovies
      tags: [movies]
      summary: Get popular movies from TMDB
      parameters:
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Popular movies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListResponse"
...
  /movies/genre/{genreId}:
    get:
      operationId: getMoviesByGenre
      tags: [movies]
      summary: Get movies by genre from TMDB
      parameters:
        - name: genreId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListResponse"
...
  /movies/{tmdbId}:
    get:
      operationId: getMovie
      tags: [movies]
      summary: Get movie details from TMDB
      parameters:
        - name: tmdbId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movie details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDetail"
  /movies/{tmdbId}/recommendations:
    get:
      operationId: getMovieRecommendations
      tags: [movies]
      summary: Get TMDB movie recommendations
      parameters:
        - name: tmdbId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListResponse"
  /movies/{tmdbId}/user-status:
    get:
      operationId: getMovieUserStatus
      tags: [movies]
      summary: Get current user's status for a movie (watchlists it belongs to, review)
      parameters:
        - name: tmdbId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieUserStatus"

  # ── Watchlists ─────────────────────────────────────────────────────────────
...
  /friends/request:
    post:
      operationId: sendFriendRequest
      tags: [friends]
      summary: Send a friend request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequestInput"
      responses:
        "201":
          description: Request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friendship"
  /friends/{id}/accept:
    patch:
      operationId: acceptFriendRequest
      tags: [friends]
      summary: Accept a friend request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friendship"
  /friends/{id}/reject:
    patch:
      operationId: rejectFriendRequest
      tags: [friends]
      summary: Reject a friend request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Friendship"
  /friends/{id}:
    delete:
      operationId: removeFriend
      tags: [friends]
      summary: Remove a friend
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /friends/movie/{tmdbId}:
    get:
      operationId: getFriendsWhoWatchedMovie
      tags: [friends]
      summary: Get friends who watched a specific movie
      parameters:
        - name: tmdbId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Friends with their ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FriendMovieActivity"

  # ── Movie Recommendations (social) ─────────────────────────────────────────
  /recommendations:
    post:
      operationId: sendRecommendation
      tags: [recommendations]
      summary: Send a movie recommendation to a friend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
...
  /recommendations/{id}/read:
    patch:
      operationId: markRecommendationRead
      tags: [recommendations]
      summary: Mark a recommendation as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"

  # ── Notifications ──────────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ── Random Pick ───────────────────────────────────────────────────────────
  /random-pick:
    get:
      operationId: getRandomPick
      tags: [movies]
      summary: Pick a random movie from watchlist or by genre
      parameters:
        - name: watchlistId
          in: query
          schema:
            type: integer
        - name: genreId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Random movie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieSummary"
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, username, displayName, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        watchedCount:
          type: integer
        reviewCount:
          type: integer
        friendCount:
          type: integer
...
    Genre:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
...
    MovieSummary:
      type: object
      required: [tmdbId, title, posterPath, releaseYear, voteAverage, popularity]
      properties:
        tmdbId:
          type: integer
        title:
          type: string
        posterPath:
          type: ["string", "null"]
        backdropPath:
          type: ["string", "null"]
        releaseYear:
          type: ["integer", "null"]
        voteAverage:
          type: number
        popularity:
          type: number
        genreIds:
          type: array
          items:
            type: integer
        overview:
          type: string
    MovieListResponse:
      type: object
      required: [results, totalPages, page]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/MovieSummary"
        totalPages:
          type: integer
        page:
          type: integer
    CastMember:
      type: object
      required: [id, name, character]
      properties:
        id:
          type: integer
        name:
          type: string
        character:
          type: string
        profilePath:
          type: ["string", "null"]
    MovieDetail:
      type: object
      required: [tmdbId, title, overview, releaseYear, voteAverage, popularity, genres, runtime]
      properties:
        tmdbId:
          type: integer
        title:
          type: string
        overview:
          type: string
        posterPath:
          type: ["string", "null"]
        backdropPath:
          type: ["string", "null"]
        releaseYear:
          type: ["integer", "null"]
        voteAverage:
          type: number
        popularity:
          type: number
        runtime:
          type: ["integer", "null"]
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
        cast:
          type: array
          items:
            $ref: "#/components/schemas/CastMember"
        trailerKey:
          type: ["string", "null"]
        dominantColor:
          type: ["string", "null"]
    MovieUserStatus:
      type: object
      required: [watchlistIds, hasReview]
      properties:
        watchlistIds:
          type: array
          items:
            type: integer
        hasReview:
          type: boolean
        rating:
          type: ["number", "null"]
...
    Friendship:
      type: object
      required: [id, requesterId, receiverId, status, createdAt]
      properties:
        id:
          type: integer
        requesterId:
          type: string
        receiverId:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        otherUser:
          $ref: "#/components/schemas/UserProfile"
    FriendRequestInput:
      type: object
      required: [targetUsername]
      properties:
        targetUsername:
          type: string
    FriendMovieActivity:
      type: object
      required: [user, rating]
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        rating:
          type: ["number", "null"]
        review:
          type: ["string", "null"]
        isSpoiler:
          type: boolean
    Recommendation:
      type: object
      required: [id, fromUserId, toUserId, tmdbId, seen, createdAt]
      properties:
        id:
          type: integer
        fromUserId:
          type: string
        toUserId:
          type: string
        tmdbId:
          type: integer
        message:
          type: ["string", "null"]
        seen:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    RecommendationInput:
      type: object
      required: [toUserId, tmdbId]
      properties:
        toUserId:
          type: string
        tmdbId:
          type: integer
        movieTitle:
          type: string
        moviePosterPath:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
    Notification:
      type: object
      required: [id, userId, type, data, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        type:
          type: string
        data:
          type: object
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time