# lib/api-zod/src/generated/types/friendship.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CineClub API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from './userProfile';

export interface Friendship {
  id: number;
  requesterId: string;
  receiverId: string;
  status: string;
  createdAt: Date;
  otherUser?: UserProfile;
}