# lib/api-zod/src/generated/types/movieSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CineClub API
 * OpenAPI spec version: 0.1.0
 */

export interface MovieSummary {
  tmdbId: number;
  title: string;
  /** @nullable */
  posterPath: string | null;
  /** @nullable */
  backdropPath?: string | null;
  /** @nullable */
  releaseYear: number | null;
  voteAverage: number;
  popularity: number;
  genreIds?: number[];
  overview?: string;
}