# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CineClub API
 * OpenAPI spec version: 0.1.0
 */

export interface Recommendation {
  id: number;
  fromUserId: string;
  toUserId: string;
  tmdbId: number;
  /** @nullable */
  message?: string | null;
  seen: boolean;
  createdAt: Date;
}