# lib/api-zod/src/generated/types/recommendationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CineClub API
 * OpenAPI spec version: 0.1.0
 */

export interface RecommendationInput {
  toUserId: string;
  tmdbId: number;
  movieTitle?: string;
  /** @nullable */
  moviePosterPath?: string | null;
  /** @nullable */
  message?: string | null;
}